gh-79315: Add Include/cpython/pythread.h header
